Updates for 64bit mode
Former-commit-id: a3ac624910264f8609628feb6521a2f006d29ab0

+++ Examples/PSPDFCatalog/Customization/PSCCustomToolbarController.m
@@ -112,7 +112,7 @@
         [indicator sizeToFit];
         [indicator startAnimating];
         indicator.tag = PSPDFLoadingViewTag;
-        indicator.frame = CGRectMake(floorf((pageView.frame.size.width - indicator.frame.size.width)/2), floorf((pageView.frame.size.height - indicator.frame.size.height)/2), indicator.frame.size.width, indicator.frame.size.height);
+        indicator.frame = CGRectMake(floor((pageView.frame.size.width - indicator.frame.size.width)/2), floor((pageView.frame.size.height - indicator.frame.size.height)/2), indicator.frame.size.width, indicator.frame.size.height);
         [pageView addSubview:indicator];
     }
 }
